Update riskdetection-get.md
small correction to clarify that specifying a GUID in this call will return a specific risk detection object, not all the risk detections for a specific user. I tested this on my demo tenant and you cannot make a call here to pivot on the userID, as far as I can tell.